Bot/requirements.txt
discord.py
Pillow
requests
python-dotenv